newsapp/requirements.txt
Django==1.4.0
MySQL-python==1.2.3
PyYAML==3.10
South==0.7.3
amqplib==1.0.2
argparse==1.2.1
celery==2.5.0
django-celery==2.5.0
django-picklefield==0.2.0
django-supervisor==0.2.7
kombu==2.1.1
requests==10.8
supervisor==3.0a12
lxml==2.3
pyquery==1.1.1
django-jsonfield==0.7.1
django-batch-select==0.2.3
